Solving some problems where the movement with bow down wasn't satisfactory
Now the movement when the bow is down is oriented to whichever direction the character is going, but when the bow is up, the movement direction is to where the camera is pointing. This is the way because there is no strafing animations for the bow down part of the character.

Therefore, the gameplay reflected is a game where you use your bow down most of the times to move through the environment and only draw it when you actually want to shoot, because you have a velocity affordance for having the bow up